refactor(ui): extract tool-status.ts and SyntaxHighlighter.tsx from Message/ToolCall/ToolsPanel
Round 5 architecture deepening — Candidates #1 + #2.

1. tool-status.ts (new ~100 lines) — unified getStatusIcon,
   getStatusColor, formatDuration, formatGitCommand, hasToolMarkers.
   Accepts both ToolStatus enum and ToolCallStatus string union via
   UnifiedToolStatus type. Eliminates 6 duplicate definitions across
   3 files.

2. SyntaxHighlighter.tsx (new ~130 lines) — SYNTAX_PATTERNS (17 regex
   patterns), getSyntaxStyle, SyntaxHighlighted memo component
   extracted from Message.tsx. Self-contained diff/git renderer with
   clean { text: string } interface.

3. ToolCall.tsx — now imports from tool-status.ts, re-exports
   ToolCallStatus type for backward compat.

4. ToolsPanel.tsx — now imports from tool-status.ts.

5. Message.tsx (382→~200 lines) — now imports from tool-status.ts and
   SyntaxHighlighter.tsx, re-exports ToolCallStatus.

All 1288 tests pass. Typecheck clean. Lint clean (239 files).
